SCRUM 85

+++ backend/src/main/java/com/greenfoxacademy/backend/models/User.java
@@ -4,7 +4,9 @@ import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
+
 import java.util.UUID;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Getter;
@@ -26,7 +28,7 @@ public abstract class User implements UserDetails {
   @Getter
   @Id
   @GeneratedValue
-  private Integer id;
+  private Long id;
   private String firstName;
   private String lastName;
   @Column(unique = true)
